+++ CHANGELOG.md
@@ -10,6 +10,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 ### Fixed
 
+## [2.0.1] 2024-09-08
+### Added
+### Changed
+### Fixed
+- display expanded text to ui console as well
+
+## [2.0.1] 2024-09-08
+### Added
+- new notification method in OMFLog allowing to expand extra text
+### Changed
+### Fixed
+
 ## [2.0.0] - 2024-09-07
 ### Added
 - IntelliJ install info in README